fix: move common functions to jira package.
