+++ pkg/tracking/jira/tickets_test.go
@@ -10,7 +10,6 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/adevinta/vulcan-tracker/pkg/common"
 	vterrors "github.com/adevinta/vulcan-tracker/pkg/errors"
 	"github.com/adevinta/vulcan-tracker/pkg/model"
 	testutil "github.com/adevinta/vulcan-tracker/pkg/testutils"
@@ -192,7 +191,7 @@ func setupSubTest(t *testing.T, transitions map[string][]model.Transition) {
 		TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 		FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 		Summary:     "Summary TEST-1",
-		Description: common.GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+		Description: GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 		Project:     "TEST",
 		Status:      ToDo,
 		TicketType:  "Vulnerability",
@@ -204,7 +203,7 @@ func setupSubTest(t *testing.T, transitions map[string][]model.Transition) {
 		TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 		FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 		Summary:     "Summary TEST-2",
-		Description: common.GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+		Description: GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 		Project:     "TEST",
 		Status:      InProgress,
 		TicketType:  "Vulnerability",
@@ -239,7 +238,7 @@ func TestGetTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-1",
-				Description: common.GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      ToDo,
 				TicketType:  "Vulnerability",
@@ -329,7 +328,7 @@ func TestCreateTicket(t *testing.T) {
 				TeamID:      "11c2c999-14a4-434f-ab02-107e2cc14324",
 				FindingID:   "fcafb55e-f8d4-4a10-b073-149b84554b94",
 				Summary:     "Summary New Ticket",
-				Description: common.GenerateDescriptionText("Description New Ticket", "fcafb55e-f8d4-4a10-b073-149b84554b94", "11c2c999-14a4-434f-ab02-107e2cc14324"),
+				Description: GenerateDescriptionText("Description New Ticket", "fcafb55e-f8d4-4a10-b073-149b84554b94", "11c2c999-14a4-434f-ab02-107e2cc14324"),
 				Project:     "TEST",
 				Status:      ToDo,
 				TicketType:  "Vulnerability",
@@ -378,7 +377,7 @@ func TestFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-1",
-				Description: common.GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -401,7 +400,7 @@ func TestFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-2",
-				Description: common.GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -422,7 +421,7 @@ func TestFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-1",
-				Description: common.GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -443,7 +442,7 @@ func TestFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-2",
-				Description: common.GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -509,7 +508,7 @@ func TestWontFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-1",
-				Description: common.GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -532,7 +531,7 @@ func TestWontFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-2",
-				Description: common.GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -553,7 +552,7 @@ func TestWontFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-1",
-				Description: common.GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-1", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
@@ -574,7 +573,7 @@ func TestWontFixTicket(t *testing.T) {
 				TeamID:      "ff9d5142-0eb3-494a-8626-a72b7182cdb2",
 				FindingID:   "4c24526d-2651-4873-8024-b27361b69723",
 				Summary:     "Summary TEST-2",
-				Description: common.GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
+				Description: GenerateDescriptionText("Description TEST-2", "4c24526d-2651-4873-8024-b27361b69723", "ff9d5142-0eb3-494a-8626-a72b7182cdb2"),
 				Project:     "TEST",
 				Status:      Resolved,
 				TicketType:  "Vulnerability",
